Rename `check/cpp_*` files to `check/cpp/*`.
