+++ toolchain/check/BUILD
@@ -21,11 +21,11 @@ cc_library(
         "context.cpp",
         "control_flow.cpp",
         "convert.cpp",
-        "cpp_custom_type_mapping.cpp",
-        "cpp_import.cpp",
-        "cpp_overload_resolution.cpp",
-        "cpp_thunk.cpp",
-        "cpp_type_mapping.cpp",
+        "cpp/custom_type_mapping.cpp",
+        "cpp/import.cpp",
+        "cpp/overload_resolution.cpp",
+        "cpp/thunk.cpp",
+        "cpp/type_mapping.cpp",
         "decl_name_stack.cpp",
         "deduce.cpp",
         "deferred_definition_worklist.cpp",
@@ -70,11 +70,11 @@ cc_library(
         "context.h",
         "control_flow.h",
         "convert.h",
-        "cpp_custom_type_mapping.h",
-        "cpp_import.h",
-        "cpp_overload_resolution.h",
-        "cpp_thunk.h",
-        "cpp_type_mapping.h",
+        "cpp/custom_type_mapping.h",
+        "cpp/import.h",
+        "cpp/overload_resolution.h",
+        "cpp/thunk.h",
+        "cpp/type_mapping.h",
         "decl_introducer_state.h",
         "decl_name_stack.h",
         "deduce.h",
