+++ toolchain/check/call.cpp
@@ -10,8 +10,8 @@
 #include "toolchain/check/context.h"
 #include "toolchain/check/control_flow.h"
 #include "toolchain/check/convert.h"
-#include "toolchain/check/cpp_overload_resolution.h"
-#include "toolchain/check/cpp_thunk.h"
+#include "toolchain/check/cpp/overload_resolution.h"
+#include "toolchain/check/cpp/thunk.h"
 #include "toolchain/check/deduce.h"
 #include "toolchain/check/facet_type.h"
 #include "toolchain/check/function.h"
