+++ toolchain/check/cpp/custom_type_mapping.h
