+++ toolchain/check/cpp/import.cpp
@@ -2,7 +2,7 @@
 // Exceptions. See /LICENSE for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "toolchain/check/cpp_import.h"
+#include "toolchain/check/cpp/import.h"
 
 #include <memory>
 #include <optional>
@@ -31,8 +31,8 @@
 #include "toolchain/check/context.h"
 #include "toolchain/check/control_flow.h"
 #include "toolchain/check/convert.h"
-#include "toolchain/check/cpp_custom_type_mapping.h"
-#include "toolchain/check/cpp_thunk.h"
+#include "toolchain/check/cpp/custom_type_mapping.h"
+#include "toolchain/check/cpp/thunk.h"
 #include "toolchain/check/diagnostic_helpers.h"
 #include "toolchain/check/eval.h"
 #include "toolchain/check/function.h"
