+++ toolchain/check/cpp/overload_resolution.cpp
@@ -2,12 +2,12 @@
 // Exceptions. See /LICENSE for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "toolchain/check/cpp_overload_resolution.h"
+#include "toolchain/check/cpp/overload_resolution.h"
 
 #include "clang/Sema/Overload.h"
 #include "clang/Sema/Sema.h"
-#include "toolchain/check/cpp_import.h"
-#include "toolchain/check/cpp_type_mapping.h"
+#include "toolchain/check/cpp/import.h"
+#include "toolchain/check/cpp/type_mapping.h"
 #include "toolchain/sem_ir/expr_info.h"
 #include "toolchain/sem_ir/typed_insts.h"
