+++ toolchain/check/type_completion.cpp
@@ -6,7 +6,7 @@
 
 #include "llvm/ADT/SmallVector.h"
 #include "toolchain/base/kind_switch.h"
-#include "toolchain/check/cpp_import.h"
+#include "toolchain/check/cpp/import.h"
 #include "toolchain/check/generic.h"
 #include "toolchain/check/inst.h"
 #include "toolchain/check/type.h"
